+++ app/src/test/java/com/srdroid/memedb/usecase/GetMemeUseCaseUT.kt
@@ -7,7 +7,7 @@ import com.srdroid.memedb.domain.errorhandler.GeneralErrorHandlerImpl
 import com.srdroid.memedb.data.model.MemeDTO
 import com.srdroid.memedb.data.repository.MemeRepositoryImpl
 import com.srdroid.memedb.domain.mappers.MemeModelMapper
-import com.srdroid.memedb.domain.use_case.GetMemeUseCase
+import com.srdroid.memedb.domain.usecases.GetMemeUseCase
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
