+++ app/src/test/java/com/srdroid/memedb/viewmodel/MemeDetailViewModelUT.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import com.srdroid.memedb.core.ID
 import com.srdroid.memedb.core.MockResponse
 import com.srdroid.memedb.core.TestCoroutineRule
-import com.srdroid.memedb.domain.use_case.GetMemeDetailsUseCase
+import com.srdroid.memedb.domain.usecases.GetMemeDetailsUseCase
 import com.srdroid.memedb.presentation.mapper.ErrorViewMapper
 import com.srdroid.memedb.presentation.mapper.MemeMapper
 import com.srdroid.memedb.presentation.memedetails.MemeDetailsViewModel
